Add HardpointBalanceForcesOnInActiveState setting
This control a new CSC feature. When set to True, the CSC will switch on
HP force balancing (offloading of the HP forces to the M1M3 force
actuators) as soon as the mirror is raised.

As this sometimes causes problems when switched on automatically too
early, this is now disabled. When the M1M3 CSC code is fixed, this can
be set to true.